SpinServer to a separate file
